requirements.txt
torch==1.13.1
torchvision==0.14.1
torchmetrics
numpy
tqdm
matplotlib
pandas
wandb
hydra_core

black
isort
pre-commit
